+++ Cookcademy.xcodeproj/xcuserdata/otaviolourenco.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "D60BAD1B-7976-4883-BB6D-95CB7A52CC9D"
+   type = "1"
+   version = "2.0">
+</Bucket>
