--- Cookcademy/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  Cookcademy
-//
-//  Created by Otavio Lourenço on 09/11/2024.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
