+++ Cookcademy/CookcademyApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct CookcademyApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            RecipesListView()
         }
     }
 }
