+++ Cookcademy/ViewModels/AppColor.swift
@@ -1,0 +1,14 @@
+//
+//  AppColor.swift
+//  Cookcademy
+//
+//  Created by Otavio Lourenço on 13/11/2024.
+//
+
+import SwiftUI
+
+struct AppColor {
+    static let background: Color = Color(.sRGB, red: 228/255, green: 235/255, blue: 250/255, opacity: 1)
+    
+    static let foreground: Color = Color(.sRGB, red: 118/255, green: 119/255, blue: 231/255, opacity: 1)
+}
